+++ capstone-backend/routes/auth.js
@@ -1,33 +1,32 @@
 const express = require("express");
 const router = express.Router();
 const Parse = require("parse/node");
-const {PARSE_APP_ID, PARSE_JAVASCRIPT_KEY} = require('./utils/config')
+const {PARSE_APP_ID, PARSE_JAVASCRIPT_KEY} = require('../utils/config')
 Parse.initialize(PARSE_APP_ID, PARSE_JAVASCRIPT_KEY);
 Parse.serverURL = "https://parseapi.back4app.com";
 
 
 router.post("/login", async (req, res) => {
-    const userRef = new Parse.Object("User");
-    Parse.User.enableUnsafeCurrentUser();
     const infoUser = req.body;
     try {
-      var currentUser = Parse.User.current();
-      if (currentUser) {
-        Parse.User.logOut();
-      }
+      // var currentUser = Parse.User.current();
+      // if (currentUser) {
+      //   Parse.User.logOut();
+      // }
       const user = await Parse.User.logIn(
         infoUser.email,
         infoUser.password,
         infoUser.username
       );
+      const sessionToken = user.getSessionToken();
       res.send({
+        sessionToken: sessionToken,
         userInfo: user,
         loginMessage: "User logged in!",
         typeStatus: "success",
         infoUser: infoUser,
       });
     } catch (error) {
-      handleErrorParse(error);
       res.send({
         loginMessage: error.message,
         typeStatus: "danger",
